+++ wp-content/plugins/core/src/Theme_Customizer/Customizer_Loader.php
@@ -2,6 +2,8 @@
 
 namespace Tribe\Project\Theme_Customizer;
 
+use WP_Customize_Manager;
+
 /**
  * @package Tribe\Project\Theme_Customizer
  */
@@ -16,7 +18,7 @@ class Customizer_Loader {
 	 *
 	 * @action customize_register
 	 */
-	public function register_customizer_controls( \WP_Customize_Manager $wp_customize ): void {
+	public function register_customizer_controls( WP_Customize_Manager $wp_customize ): void {
 	}
 
 }